monorepo/apps/webscraper/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
httpx==0.28.1
beautifulsoup4==4.14.3
playwright==1.58.0
pydantic-settings==2.13.1
python-dotenv==1.2.2
lxml==6.0.2

# Test dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.23.1
